Move FlowExpressions to a new package, org.checkerframework.dataflow.expression
