+++ checker/src/main/java/org/checkerframework/checker/index/OffsetDependentTypesHelper.java
@@ -3,7 +3,8 @@ package org.checkerframework.checker.index;
 import com.sun.source.tree.MemberSelectTree;
 import com.sun.source.util.TreePath;
 import org.checkerframework.checker.index.upperbound.OffsetEquation;
-import org.checkerframework.dataflow.analysis.FlowExpressions;
+import org.checkerframework.dataflow.expression.FieldAccess;
+import org.checkerframework.dataflow.expression.Receiver;
 import org.checkerframework.framework.type.AnnotatedTypeFactory;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.framework.type.treeannotator.TreeAnnotator;
@@ -34,7 +35,7 @@ public class OffsetDependentTypesHelper extends DependentTypesHelper {
         }
         if (expression.indexOf('-') == -1 && expression.indexOf('+') == -1) {
             // The expression contains no "-" or "+", so it can be standardized directly.
-            FlowExpressions.Receiver result;
+            Receiver result;
             try {
                 result =
                         FlowExpressionParseUtil.parse(
@@ -45,10 +46,8 @@ public class OffsetDependentTypesHelper extends DependentTypesHelper {
             if (result == null) {
                 return new DependentTypesError(expression, " ").toString();
             }
-            if (result instanceof FlowExpressions.FieldAccess
-                    && ((FlowExpressions.FieldAccess) result).isFinal()) {
-                Object constant =
-                        ((FlowExpressions.FieldAccess) result).getField().getConstantValue();
+            if (result instanceof FieldAccess && ((FieldAccess) result).isFinal()) {
+                Object constant = ((FieldAccess) result).getField().getConstantValue();
                 if (constant != null && !(constant instanceof String)) {
                     return constant.toString();
                 }
