+++ checker/src/main/java/org/checkerframework/checker/lock/LockTransfer.java
@@ -7,8 +7,6 @@ import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.Modifier;
 import javax.lang.model.type.TypeMirror;
-import org.checkerframework.dataflow.analysis.FlowExpressions;
-import org.checkerframework.dataflow.analysis.FlowExpressions.Receiver;
 import org.checkerframework.dataflow.analysis.TransferInput;
 import org.checkerframework.dataflow.analysis.TransferResult;
 import org.checkerframework.dataflow.cfg.UnderlyingAST;
@@ -17,6 +15,9 @@ import org.checkerframework.dataflow.cfg.UnderlyingAST.Kind;
 import org.checkerframework.dataflow.cfg.node.LocalVariableNode;
 import org.checkerframework.dataflow.cfg.node.Node;
 import org.checkerframework.dataflow.cfg.node.SynchronizedNode;
+import org.checkerframework.dataflow.expression.ClassName;
+import org.checkerframework.dataflow.expression.FlowExpressions;
+import org.checkerframework.dataflow.expression.Receiver;
 import org.checkerframework.framework.flow.CFAbstractTransfer;
 import org.checkerframework.framework.flow.CFValue;
 import org.checkerframework.javacutil.TreeUtils;
@@ -112,8 +113,7 @@ public class LockTransfer extends CFAbstractTransfer<CFValue, LockStore, LockTra
                 TypeMirror classType = TreeUtils.typeOf(classTree);
 
                 if (methodElement.getModifiers().contains(Modifier.STATIC)) {
-                    store.insertValue(
-                            new FlowExpressions.ClassName(classType), atypeFactory.LOCKHELD);
+                    store.insertValue(new ClassName(classType), atypeFactory.LOCKHELD);
                 } else {
                     store.insertThisValue(atypeFactory.LOCKHELD, classType);
                 }
