+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/StringCFGVisualizer.java
@@ -10,13 +10,17 @@ import org.checkerframework.checker.nullness.qual.KeyFor;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.dataflow.analysis.AbstractValue;
 import org.checkerframework.dataflow.analysis.Analysis;
-import org.checkerframework.dataflow.analysis.FlowExpressions;
 import org.checkerframework.dataflow.analysis.Store;
 import org.checkerframework.dataflow.analysis.TransferFunction;
 import org.checkerframework.dataflow.cfg.AbstractCFGVisualizer.VisualizeWhere;
 import org.checkerframework.dataflow.cfg.block.Block;
 import org.checkerframework.dataflow.cfg.block.ConditionalBlock;
 import org.checkerframework.dataflow.cfg.block.SpecialBlock;
+import org.checkerframework.dataflow.expression.ArrayAccess;
+import org.checkerframework.dataflow.expression.ClassName;
+import org.checkerframework.dataflow.expression.FieldAccess;
+import org.checkerframework.dataflow.expression.LocalVariable;
+import org.checkerframework.dataflow.expression.MethodCall;
 
 /** Generate the String representation of a control flow graph. */
 public class StringCFGVisualizer<
@@ -106,27 +110,27 @@ public class StringCFGVisualizer<
     }
 
     @Override
-    public String visualizeStoreLocalVar(FlowExpressions.LocalVariable localVar, V value) {
+    public String visualizeStoreLocalVar(LocalVariable localVar, V value) {
         return storeEntryIndent + localVar + " > " + value;
     }
 
     @Override
-    public String visualizeStoreFieldVal(FlowExpressions.FieldAccess fieldAccess, V value) {
+    public String visualizeStoreFieldVal(FieldAccess fieldAccess, V value) {
         return storeEntryIndent + fieldAccess + " > " + value;
     }
 
     @Override
-    public String visualizeStoreArrayVal(FlowExpressions.ArrayAccess arrayValue, V value) {
+    public String visualizeStoreArrayVal(ArrayAccess arrayValue, V value) {
         return storeEntryIndent + arrayValue + " > " + value;
     }
 
     @Override
-    public String visualizeStoreMethodVals(FlowExpressions.MethodCall methodCall, V value) {
+    public String visualizeStoreMethodVals(MethodCall methodCall, V value) {
         return storeEntryIndent + methodCall + " > " + value;
     }
 
     @Override
-    public String visualizeStoreClassVals(FlowExpressions.ClassName className, V value) {
+    public String visualizeStoreClassVals(ClassName className, V value) {
         return storeEntryIndent + className + " > " + value;
     }
