+++ dataflow/src/main/java/org/checkerframework/dataflow/livevariable/LiveVarStore.java
@@ -4,7 +4,6 @@ import java.util.HashSet;
 import java.util.Set;
 import java.util.StringJoiner;
 import org.checkerframework.checker.nullness.qual.Nullable;
-import org.checkerframework.dataflow.analysis.FlowExpressions.Receiver;
 import org.checkerframework.dataflow.analysis.Store;
 import org.checkerframework.dataflow.cfg.CFGVisualizer;
 import org.checkerframework.dataflow.cfg.node.BinaryOperationNode;
@@ -15,6 +14,7 @@ import org.checkerframework.dataflow.cfg.node.Node;
 import org.checkerframework.dataflow.cfg.node.TernaryExpressionNode;
 import org.checkerframework.dataflow.cfg.node.TypeCastNode;
 import org.checkerframework.dataflow.cfg.node.UnaryOperationNode;
+import org.checkerframework.dataflow.expression.Receiver;
 import org.checkerframework.javacutil.BugInCF;
 
 /** A live variable store contains a set of live variables represented by nodes. */
